pyproject.toml
[tool.poetry]
name = "rush-cli"
version = "0.2.0"
description = "Minimalistic bash task runner"
authors = ["rednafi <redowan.nafi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
colorama = "^0.4.3"
pyyaml = "^5.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
rush = 'rush_cli.cli:entrypoint'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
